take component/IconsWrapper apart into each components/icons/xxx.tsx
