+++ src/pages/posts/[id].tsx
@@ -4,7 +4,7 @@ import { GetStaticProps, GetStaticPaths } from 'next'
 import { Layout } from '../../components/Layout'
 import { getAllPostIds, getPostData } from '../../lib/posts'
 // import { getFetchPath } from '../../components/HeaderImg'
-import { PostIcons } from '../../components/IconsWrapper'
+import { PostIcons } from '../../components/icons/index'
 import { Date } from '../../components/common/Date'
 import { CustomImg } from '../../components/common/Image'
 import { CustomHead } from '../../components/common/Head'
