+++ src/pages/posts/index.tsx
@@ -3,7 +3,7 @@ import css from 'styled-jsx/css'
 import { Layout } from '../../components/Layout'
 import { CustomImg } from '../../components/common/Image'
 import { getSortedPostsData } from '../../lib/posts'
-import { PostsIcons } from '../../components/IconsWrapper'
+import { PostsIcons } from '../../components/icons/index'
 import { Date } from '../../components/common/Date'
 import { GetStaticProps } from 'next'
 import { CustomHead } from '../../components/common/Head'
